refactor(cleaner): move into streams-bootstrap module
