+++ kpops/components/streams_bootstrap/streams/streams_app.py
@@ -6,11 +6,11 @@ from typing_extensions import override
 
 from kpops.component_handlers.kubernetes.pvc_handler import PVCHandler
 from kpops.components.base_components.helm_app import HelmApp
-from kpops.components.base_components.kafka_app import KafkaAppCleaner
 from kpops.components.common.app_type import AppType
 from kpops.components.common.topic import KafkaTopic
 from kpops.components.streams_bootstrap.base import (
     StreamsBootstrap,
+    StreamsBootstrapCleaner,
 )
 from kpops.components.streams_bootstrap.streams.model import (
     StreamsAppValues,
@@ -26,7 +26,7 @@ from kpops.utils.docstring import describe_attr
 log = logging.getLogger("StreamsApp")
 
 
-class StreamsAppCleaner(KafkaAppCleaner, StreamsBootstrap):
+class StreamsAppCleaner(StreamsBootstrapCleaner, StreamsBootstrap):
     from_: None = None  # pyright: ignore[reportIncompatibleVariableOverride]
     to: None = None  # pyright: ignore[reportIncompatibleVariableOverride]
     values: StreamsAppValues  # pyright: ignore[reportIncompatibleVariableOverride]
